feat: create reusable position validation
